+++ concordium-cis2/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Update `concordium-std` to version 5.
 - Add `From` implementation from types implementing `From<UpgradeError>`, `From<QueryAccountBalanceError>` or `From<QueryContractBalanceError>` to `Cis2Error`.
+- Add SchemaType for Cis2Event<T, A>
 
 ## concordium-cis2 1.2.0 (2022-09-01)